--- a/internal/processing/Token.go
+++ b/internal/processing/Token.go
@@ -3,5 +3,4 @@ package processing
 type Token struct {
 	Type TokenType
 	Text string
-	Pos  int
 }