… test/cmd.test
… test/cpu.out
… test/cpu.svg