Work on micro-optimization of the code. \
A slight improvement was obtained in terms of:
category| % | before | after 
------|:--------:|:------:|:---------:
memory used  | 16.7 | 5576091 B/op | 4644179 B/op
in time | 22.2 | 3718841 ns/op | 2891616 ns/op
in the number of allocations | 17.5 | 63467 allocs/op | 52378 allocs/op